Replace client
Replaced the `create-react-app` client with one created from scratch.
