+++ .gitignore
@@ -1,0 +1,2 @@
+
+client/node_modules/
