+++ client/src/App.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+
+function App() {
+  return (
+    <div className="App">
+      Seen
+    </div>
+  );
+}
+
+export default App;
