--- seen-client/src/App.js
@@ -1,24 +1,0 @@
-
-import React from 'react';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn Reactss
-        </a>
-      </header>
-    </div>
-  );
-}
-
-export default App;
